requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
gunicorn==23.0.0
idna==3.10
packaging==25.0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
